Initial Commit
The project checks the vowels array against the input string. 
The loop then iterates through the arrays and checks against the characters for a match. If a match is detected the character is pushed into an empty resultArray. Additionally if either an e or a u is detected in the string then an additional character of that value is added to the resultArray. Finally at the end the completed array is output to the console, the characters are joined to a single string and converted to upper case.